Allow passing a sub to 'token' attribute during construction
Up until now, the only way to authenticate against a kubernetes cluster
via a token was specifying the token (a string) on Kubectl::CLIWrapper
instantiation.

With this changeset, 'token' attribute can either be:
a) A string
b) A reference to a sub
When a sub is specified, it will be called by Kubectl::CLIWrapper during
object construction.

The intent behind this changeset is enabling token auto-refresh.